Ensure CATALINA_PID takes effect for instances on RHEL
Under RHEL 7, /usr/sbin/tomcat-sysd sets CATALINA_PID for the given
instance but it is erroneously replaced with the default value from
/etc/tomcat/tomcat.conf. The only way to work around this is to also
set it in the instance's sysconfig file.

This doesn't change the fact that /var/run is only writeable by root.
Any new PID files need to be put in place beforehand as the script
does not set the correct permissions for us. It actually says it does
in the comments… but it doesn't! This is fixed here with a file
resource.

Recent Fedora releases use JSVC to launch Tomcat instead so all this
will most likely change in RHEL 8.